use prefix rather than base_prefix for PYTHONHOME
prefix gets modified to point to a virtualenv, whereas base_prefix is
not, so required libraries will not be found when running in a
virtualenv.

Setting both prefix and exec_prefix potentially helps on systems where
architecture-specific and architecture-independent components of the
python library are split. This matches the docs:

https://docs.python.org/3/using/cmdline.html#envvar-PYTHONHOME