avoid deleting python sample type objects without GIL
The sample type lookup table holds references to python objects, and is
constructed statically. This is dangerous during initialisation (as the
python interpreter is not yet running, but actually causes problems at
finalisation.

If the python interpreter gets shut down first (which can happen as it's
static too), a pybind11 assertion fires when it's finalised, as the GIL
is not held.

Use reference counting from the callback lambdas instead.