append to PYTHONPATH in tests if possible
As per the TODO, it is a good idea to pass through PYTHONPATH from the
environment to the tests, as this is a valid way to get python to find
modules.

It's tricky to do this in older cmake versions, but this should only
really happen on old versions of ubuntu, where you're unlikely to need
it anyway.