simplify PYTHONPATH setup in libpythoncomponents tests
Rather than passing in one required path with a definition, prefer to
set it in the environment with cmake.

Python path modification seems to have been duplicated between a fixture
and the WrapUsePYTHONPATH test, which is otherwise a duplicate of
WrapMultiFilePackage.

This is simpler, and extending rather than overriding the path (as in
the fixture) makes it work in more situations.